+	* configure.ac: 0.23.2 +	Compiles and works on AIX. + +	* tests/param-unknown.sh: add -u to check. +	along with --unknown-parameter. + +	* dsh.c (do_echoing_back_process): use size_t for getline argument. + +	* configure.ac: 0.23.1.aix.2 + +	* parameter.c (read_machinenetgroup): +	work around AIX and work around void function. + +	* configure.ac: check for setnetgrent +	(SETNETGRENT_RETURNS_VOID): check for setnetgrent which +	returns void (AIX is possibly broken in this respect, so +	work around it). +

+++ dsh.c
@@ -114,7 +114,7 @@ static int
 do_echoing_back_process(int fd_in, int fd_out, const char * prompt)
 {
   char * buf = NULL;
-  int bufsize = 0;
+  size_t bufsize = 0;
   FILE*f = fdopen (fd_in, "r");
   FILE*standard_output = fdopen (fd_out, "w");
